+++ src/com/devstackio/fullstackcoder/gamestates/Game.java
@@ -4,21 +4,18 @@ import org.newdawn.slick.Color;
 import org.newdawn.slick.GameContainer;
 import org.newdawn.slick.Graphics;
 import org.newdawn.slick.SlickException;
-import org.newdawn.slick.state.BasicGameState;
 import org.newdawn.slick.state.StateBasedGame;
 
-public class Game extends BasicGameState {
-    
-    private StateBasedGame game;
+public class Game extends SharedGameState {
 
     @Override
     public int getID() {
         return 3;
     }
-
+    
     @Override
     public void init(GameContainer gc, StateBasedGame sbg) throws SlickException {
-        this.game = sbg;
+        
     }
 
     @Override